# agents/roles/security/sbom-generator/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AionRs SBOM Generator
  version: 1.0.0
  description: |
    Syft-wrapped SBOM generation service. Produces CycloneDX
    (JSON/XML) and SPDX (JSON/tag-value) for Docker images, git
    repositories, filesystems, archives, and registry catalogs.
  contact:
    name: AionRs Security Team
    email: security@aionrs.io
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:5001
    description: Local development
  - url: http://sbom-generator.aionrs.svc.cluster.local:5001
    description: Kubernetes service
...
paths:
...
  /v1/sbom/generate:
    post:
      tags: [sbom]
      summary: Generate one or more SBOMs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Generated SBOM(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Syft execution error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: Timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sbom/analyze:
    post:
      tags: [sbom]
      summary: Compatibility alias of /generate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Generated SBOM(s)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
...
components:
  schemas:
    SourceType:
      type: string
      enum: [directory, file, docker-image, oci-image, git-repository, archive, registry]
    SBOMFormat:
      type: string
      enum: [cyclonedx-json, cyclonedx-xml, spdx-json, spdx-tag-value, syft-json]
    SourceRef:
      type: object
      required: [type, value]
      properties:
        type: { $ref: '#/components/schemas/SourceType' }
        value: { type: string }
        options:
          type: object
          additionalProperties: true
    GenerateRequest:
      type: object
      required: [source]
      properties:
        source: { $ref: '#/components/schemas/SourceRef' }
        formats:
          type: array
          items: { $ref: '#/components/schemas/SBOMFormat' }
          minItems: 1
        scan_root: { type: string, nullable: true }
        exclude_paths:
          type: array
          items: { type: string }
        include_dev_dependencies: { type: boolean }
        catalogs:
          type: array
          items: { type: string }
        output_relationships: { type: boolean }
        metadata:
          type: object
          additionalProperties: { type: string }
        tenant_id: { type: string, nullable: true }
    FormattedSBOM:
      type: object
      required: [format, media_type, body, byte_size]
      properties:
        format: { $ref: '#/components/schemas/SBOMFormat' }
        media_type: { type: string }
        body: { type: string }
        byte_size: { type: integer }
    GenerateResponse:
      type: object
      properties:
        request_id: { type: string }
        job_id: { type: string }
        accepted_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, nullable: true }
        source_type: { type: string }
        source_value: { type: string }
        format: { $ref: '#/components/schemas/SBOMFormat' }
        components_count: { type: integer }
        distinct_licenses: { type: integer }
        formats:
          type: array
          items: { $ref: '#/components/schemas/FormattedSBOM' }
        warnings:
          type: array
          items: { type: string }
        bus_event_id: { type: string, nullable: true }
...
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
        request_id: { type: string, nullable: true }